refactor(sdk): single-source the recognize phantom guard (#331 follow-up)
The phantom-guard discriminator existed as two independent byte-identical copies
— inline in makeOnboarding.recognize() (SDK) and as classifyRecognition() in base's
signing-key-glue. For a security-critical "don't adopt a phantom identity" check,
two copies are a drift risk (harden one, the other silently stays vulnerable).

Promote classifyRecognition({userHandle, derivedAddress}) into gc-derived-identity.mjs
(next to makeDerivedIdentity), unit-tested once. onb.recognize() and base's glue both
import it; base re-exports it so its page glue + tests are unchanged. The now-unused
decodeAddress imports drop from both call sites. Re-vendored; SDK 0.10.0 → 0.11.0.

Surfaced by the hub-you review on #331. The deeper structural fix (a shared
recognize→{key,credentialId,verdict} primitive so key-needing consumers stop
re-implementing derive/recognize) is tracked in #332.
